trigger publish extractors

+++ tik_manager4/objects/work.py
@@ -437,7 +437,7 @@ class Work(Settings, Entity):
         thumbnail_name = f"{self._name}_v{version_number:03d}_thumbnail.jpg"
         return version_number, version_name, thumbnail_name
 
-    def load_version(self, version_number, force=False):
+    def load_version(self, version_number, force=False, **kwargs):
         """Load the given version of the work."""
         version_obj = self.get_version(version_number)
         if version_obj:
@@ -574,7 +574,8 @@ class Work(Settings, Entity):
             dest_path = self.get_purgatory_project_path(relative_path)
             # shutil.move(abs_path, dest_path, copy_function=shutil.copytree)
             Path(dest_path).parent.mkdir(parents=True, exist_ok=True)
-            shutil.move(abs_path, dest_path)
+            if Path(abs_path).exists():
+                shutil.move(abs_path, dest_path)
 
             # move the thumbnail
             thumbnail_relative_path = version_obj.get("thumbnail", None)
@@ -582,7 +583,8 @@ class Work(Settings, Entity):
                 thumbnail_abs_path = self.get_abs_database_path(thumbnail_relative_path)
                 thumbnail_dest_path = self.get_purgatory_database_path(thumbnail_relative_path)
                 Path(thumbnail_dest_path).parent.mkdir(parents=True, exist_ok=True)
-                shutil.move(thumbnail_abs_path, thumbnail_dest_path, copy_function=shutil.copytree)
+                if Path(thumbnail_abs_path).exists():
+                    shutil.move(thumbnail_abs_path, thumbnail_dest_path, copy_function=shutil.copytree)
 
             # remove the version from the versions list
             self._versions.remove(version_obj)
